# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyCarQR API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/payment-settings:
    put:
      operationId: updatePaymentSettings
      tags: [admin]
      summary: Admin — update payment settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
  /admin/payment-requests:
    get:
      operationId: getAdminPaymentRequests
      tags: [admin]
      summary: Admin — list all payment requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
  /admin/payment-requests/{requestId}/approve:
    post:
      operationId: approvePaymentRequest
      tags: [admin]
      summary: Admin — approve payment, grant premium
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                adminNote:
                  type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /admin/payment-requests/{requestId}/reject:
    post:
      operationId: rejectPaymentRequest
      tags: [admin]
      summary: Admin — reject payment request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                adminNote:
                  type: string
              required:
                - adminNote
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /admin/vehicles:
    get:
      operationId: getAdminVehicles
      tags: [admin]
      summary: Admin — list all vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle list
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminVehicle"
                  total:
                    type: integer
                required:
                  - vehicles
                  - total
  /admin/sos-profiles:
    get:
      operationId: getAdminSosProfiles
      tags: [admin]
      summary: Admin — list all SOS profiles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SOS profiles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SosProfile"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Admin overall stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersList"
  /admin/users/{userId}/upgrade:
    post:
      operationId: adminUpgradeUser
      tags: [admin]
      summary: Upgrade user to premium (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan:
                  type: string
                  enum: [free, premium]
                durationDays:
                  type: integer
              required: [plan]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/alerts:
    get:
      operationId: getAdminAlerts
      tags: [admin]
      summary: List all scan alerts (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanAlert"
                  total:
                    type: integer
                required: [alerts, total]
  /admin/accident-reports:
    get:
      operationId: getAdminAccidentReports
      tags: [admin]
      summary: List all accident reports (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Accident reports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/AccidentReport"
                  total:
                    type: integer
                required: [reports, total]
  /admin/lost-items:
    get:
      operationId: getAdminLostItems
      tags: [admin]
      summary: List all lost item reports (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Lost items list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/LostItem"
                  total:
                    type: integer
                required: [items, total]
...
  /admin/sticker-orders:
    get:
      operationId: getAdminStickerOrders
      tags: [admin]
      summary: Get all sticker orders (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: paymentStatus
          in: query
          schema:
            type: string
        - name: orderStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/StickerOrder"
                  total:
                    type: integer
                required: [orders, total]
  /admin/sticker-orders/{orderId}:
    put:
      operationId: updateStickerOrder
      tags: [admin]
      summary: Update sticker order status/tracking (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStickerOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerOrder"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ScanAlert:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        alertType:
          type: string
        message:
          type: string
        scannerLocation:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vehicleId
        - alertType
        - isRead
        - createdAt
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        premiumUsers:
          type: integer
        totalVehicles:
          type: integer
        totalAlerts:
          type: integer
        alertsToday:
          type: integer
        totalScans:
          type: integer
        activeSubscriptions:
          type: integer
        pendingPayments:
          type: integer
        accidentReports:
          type: integer
        lostItemReports:
          type: integer
      required:
        - totalUsers
        - premiumUsers
        - totalVehicles
        - totalAlerts
        - alertsToday
        - totalScans
        - activeSubscriptions
        - pendingPayments
        - accidentReports
        - lostItemReports
    AdminUser:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, premium]
        vehicleCount:
          type: integer
        alertCount:
          type: integer
        createdAt:
          type: string
        isAdmin:
          type: boolean
        premiumExpiresAt:
          type: string
          nullable: true
      required:
        - userId
        - plan
        - vehicleCount
        - alertCount
        - isAdmin
    AdminUsersList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required:
        - users
        - total
    SosProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        emergencyContactName:
          type: string
        emergencyPhone:
          type: string
        bloodGroup:
          type: string
        medicalNotes:
          type: string
        altContactName:
          type: string
        altContactPhone:
          type: string
        isEnabled:
          type: boolean
        updatedAt:
          type: string
      required:
        - id
        - userId
        - emergencyContactName
        - emergencyPhone
        - bloodGroup
        - medicalNotes
        - altContactName
        - altContactPhone
        - isEnabled
        - updatedAt
...
    AccidentReport:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        reportedAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - vehicleId
        - description
        - photos
        - reportedAt
        - isRead
...
    LostItem:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        finderContact:
          type: string
        reportedAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - vehicleId
        - message
        - photos
        - reportedAt
        - isRead
...
    PaymentSettings:
      type: object
      properties:
        id:
          type: integer
        upiId:
          type: string
        qrImageBase64:
          type: string
        monthlyPrice:
          type: integer
        yearlyPrice:
          type: integer
        instructions:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - upiId
        - qrImageBase64
        - monthlyPrice
        - yearlyPrice
        - instructions
        - updatedAt
    UpdatePaymentSettingsBody:
      type: object
      properties:
        upiId:
          type: string
        qrImageBase64:
          type: string
        monthlyPrice:
          type: integer
        yearlyPrice:
          type: integer
        instructions:
          type: string
      required:
        - upiId
        - monthlyPrice
        - yearlyPrice
        - instructions
    PaymentRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        email:
          type: string
        status:
          type: string
        planType:
          type: string
        amount:
          type: integer
        screenshotBase64:
          type: string
        adminNote:
          type: string
        createdAt:
          type: string
        reviewedAt:
          type: string
        expiresAt:
          type: string
      required:
        - id
        - userId
        - status
        - planType
        - amount
        - createdAt
...
    AdminVehicle:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        email:
          type: string
        vehicleNumber:
          type: string
        brand:
          type: string
        model:
          type: string
        vehicleType:
          type: string
        qrCode:
          type: string
          nullable: true
        qrActive:
          type: boolean
        plan:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - vehicleNumber
        - qrActive
        - createdAt
    StickerOrder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        vehicleId:
          type: integer
          nullable: true
        vehicleQrCode:
          type: string
          nullable: true
        vehicleNumber:
          type: string
          nullable: true
        customerName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        product:
          type: string
        amount:
          type: integer
        stickerStyle:
          type: string
          nullable: true
        screenshotBase64:
          type: string
          nullable: true
        paymentStatus:
          type: string
        orderStatus:
          type: string
        trackingNumber:
          type: string
          nullable: true
        adminNote:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - customerName
        - phone
        - addressLine1
        - city
        - state
        - pincode
        - product
        - amount
        - paymentStatus
        - orderStatus
        - createdAt
        - updatedAt
...
    UpdateStickerOrderRequest:
      type: object
      properties:
        paymentStatus:
          type: string
        orderStatus:
          type: string
        trackingNumber:
          type: string
        adminNote:
          type: string